+++ src/any_vec.rs
@@ -2,7 +2,7 @@ use std::alloc::Layout;
 use std::any::TypeId;
 use std::marker::PhantomData;
 use std::mem::ManuallyDrop;
-use std::ops::{Range, RangeBounds};
+use std::ops::{Deref, Index, Range, RangeBounds};
 use std::ptr::NonNull;
 use std::slice;
 use crate::{AnyVecTyped, into_range, ops, refs};
@@ -12,7 +12,7 @@ use crate::ops::{TempValue, Remove, SwapRemove, remove, swap_remove};
 use crate::ops::{Drain, Splice, drain, splice};
 use crate::any_vec::traits::{None};
 use crate::clone_type::{CloneFn, CloneFnTrait, CloneType};
-use crate::element::{AnyElement, ElementMut, ElementRef};
+use crate::element::{ElementPointer, Element, ElementMut, ElementRef};
 use crate::any_vec_ptr::AnyVecPtr;
 use crate::iter::{Iter, IterMut, IterRef};
 use crate::traits::{Cloneable, Trait};
@@ -155,7 +155,7 @@ impl<Traits: ?Sized + Trait> AnyVec<Traits>
 
     #[inline]
     pub unsafe fn downcast_ref_unchecked<Element: 'static>(&self) -> AnyVecRef<Element> {
-        refs::Ref(AnyVecTyped::new(NonNull::from(&self.raw)))
+        AnyVecRef(AnyVecTyped::new(NonNull::from(&self.raw)))
     }
 
     #[inline]
@@ -189,11 +189,11 @@ impl<Traits: ?Sized + Trait> AnyVec<Traits>
     }
 
     #[inline]
-    unsafe fn get_element(&self, index: usize) -> ManuallyDrop<AnyElement<AnyVecPtr<Traits>>>{
+    unsafe fn get_element(&self, index: usize) -> ManuallyDrop<ElementPointer<AnyVecPtr<Traits>>>{
         let element = NonNull::new_unchecked(
             self.as_bytes().add(self.element_layout().size() * index) as *mut u8
         );
-        ManuallyDrop::new(AnyElement::new(
+        ManuallyDrop::new(ElementPointer::new(
             AnyVecPtr::from(self),
             element
         ))
@@ -426,8 +426,22 @@ impl<'a, Traits: ?Sized + Trait> IntoIterator for &'a mut AnyVec<Traits>{
 ///
 /// [`AnyVec`]: crate::AnyVec
 /// [`AnyVec::downcast_ref`]: crate::AnyVec::downcast_ref
-pub type AnyVecRef<'a, T> = refs::Ref<AnyVecTyped<'a, T>>;
+//pub type AnyVecRef<'a, T> = refs::Ref<AnyVecTyped<'a, T>>;
+pub struct AnyVecRef<'a, T: 'static>(pub(crate) AnyVecTyped<'a, T>);
+impl<'a, T: 'static> Clone for AnyVecRef<'a, T>{
+    #[inline]
+    fn clone(&self) -> Self {
+        Self(self.0.clone())
+    }
+}
+impl<'a, T: 'static> Deref for AnyVecRef<'a, T>{
+    type Target = AnyVecTyped<'a, T>;
 
+    #[inline]
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
 impl<'a, T: 'static> IntoIterator for AnyVecRef<'a, T>{
     type Item = &'a T;
     type IntoIter = slice::Iter<'a, T>;