+++ src/any_vec_typed.rs
@@ -39,6 +39,12 @@ impl<'a, T: 'static> AnyVecTyped<'a, T>{
         Self{any_vec, phantom: PhantomData}
     }
 
+    /// AnyVecTyped should not be Clone, because it can be both & and &mut.
+    #[inline]
+    pub(crate) fn clone(&self) -> Self {
+        Self{any_vec: self.any_vec, phantom: PhantomData}
+    }
+
     #[inline]
     fn this(&self) -> &'a AnyVecRaw {
         unsafe{ self.any_vec.as_ref() }
@@ -186,18 +192,21 @@ impl<'a, T: 'static> AnyVecTyped<'a, T>{
     }
 }
 
+// Do not implement Index, since we can't do the same for AnyVec
+/*
 impl<'a, T: 'static, I: SliceIndex<[T]>> Index<I> for AnyVecTyped<'a, T> {
     type Output = I::Output;
 
     #[inline]
-    fn index(&self, index: I) -> &Self::Output {
+    fn index(&self, index: I) -> &'a Self::Output {
         self.as_slice().index(index)
     }
 }
 
 impl<'a, T: 'static, I: SliceIndex<[T]>> IndexMut<I> for AnyVecTyped<'a, T> {
     #[inline]
-    fn index_mut(&mut self, index: I) -> &mut Self::Output {
+    fn index_mut(&mut self, index: I) -> &'a mut Self::Output {
         self.as_mut_slice().index_mut(index)
     }
 }
+ */