+++ src/element.rs
@@ -11,8 +11,11 @@ use crate::traits::{Cloneable, None, Trait};
 // Typed operations will never use type-erased Element, so there is no
 // need in type-known-based optimizations.
 
-// TODO: hide
+// TODO: hide. Derive AnyValue staff in Element.
+//
 /// Owning pointer to [`AnyVec`] element.
+///
+/// This is public, just so you can see what [`Element`] can do.
 ///
 /// # Notes
 ///
@@ -22,19 +25,19 @@ use crate::traits::{Cloneable, None, Trait};
 ///
 /// [`AnyVec`]: crate::AnyVec
 /// [`AnyVec::get`]: crate::AnyVec::get
-pub struct AnyElement<'a, AnyVecPtr: IAnyVecRawPtr>{
+pub struct ElementPointer<'a, AnyVecPtr: IAnyVecRawPtr>{
     any_vec_ptr: AnyVecPtr,
     element: NonNull<u8>,
     phantom: PhantomData<&'a mut AnyVecRaw>
 }
 
-impl<'a, AnyVecPtr: IAnyVecRawPtr> AnyElement<'a, AnyVecPtr>{
+impl<'a, AnyVecPtr: IAnyVecRawPtr> ElementPointer<'a, AnyVecPtr>{
     #[inline]
     pub(crate) fn new(any_vec_ptr: AnyVecPtr, element: NonNull<u8>) -> Self {
         Self{any_vec_ptr, element, phantom: PhantomData}
     }
 
-    /// AnyElement should not be `Clone`.
+    /// ElementPointer should not be `Clone`, because it can be both & and &mut.
     #[inline]
     pub(crate) fn clone(&self) -> Self {
         Self::new(self.any_vec_ptr, self.element)
@@ -74,7 +77,7 @@ impl<'a, AnyVecPtr: IAnyVecRawPtr> AnyElement<'a, AnyVecPtr>{
     }
 }
 
-impl<'a, AnyVecPtr: IAnyVecRawPtr> Drop for AnyElement<'a, AnyVecPtr>{
+impl<'a, AnyVecPtr: IAnyVecRawPtr> Drop for ElementPointer<'a, AnyVecPtr>{
     #[inline]
     fn drop(&mut self) {
         if let Some(drop_fn) = self.any_vec_raw().drop_fn(){
@@ -83,7 +86,7 @@ impl<'a, AnyVecPtr: IAnyVecRawPtr> Drop for AnyElement<'a, AnyVecPtr>{
     }
 }
 
-impl<'a, AnyVecPtr: IAnyVecRawPtr> AnyValue for AnyElement<'a, AnyVecPtr>{
+impl<'a, AnyVecPtr: IAnyVecRawPtr> AnyValue for ElementPointer<'a, AnyVecPtr>{
     type Type = AnyVecPtr::Element;
 
     #[inline]
@@ -102,10 +105,10 @@ impl<'a, AnyVecPtr: IAnyVecRawPtr> AnyValue for AnyElement<'a, AnyVecPtr>{
     }
 }
 
-impl<'a, AnyVecPtr: IAnyVecRawPtr> AnyValueMut for AnyElement<'a, AnyVecPtr>{}
+impl<'a, AnyVecPtr: IAnyVecRawPtr> AnyValueMut for ElementPointer<'a, AnyVecPtr>{}
 
 impl<'a, Traits: ?Sized + Cloneable + Trait>
-    AnyValueCloneable for AnyElement<'a, AnyVecPtr<Traits>>
+    AnyValueCloneable for ElementPointer<'a, AnyVecPtr<Traits>>
 {
     #[inline]
     unsafe fn clone_into(&self, out: *mut u8) {
@@ -113,14 +116,14 @@ impl<'a, Traits: ?Sized + Cloneable + Trait>
     }
 }
 
-unsafe impl<'a, Traits: ?Sized + Send + Trait> Send for AnyElement<'a, AnyVecPtr<Traits>>{}
-unsafe impl<'a, Traits: ?Sized + Sync + Trait> Sync for AnyElement<'a, AnyVecPtr<Traits>>{}
+unsafe impl<'a, Traits: ?Sized + Send + Trait> Send for ElementPointer<'a, AnyVecPtr<Traits>>{}
+unsafe impl<'a, Traits: ?Sized + Sync + Trait> Sync for ElementPointer<'a, AnyVecPtr<Traits>>{}
 
 
 /// [`AnyVec`] element.
 ///
 /// [`AnyVec`]: crate::AnyVec
-pub type Element<'a, Traits> = AnyElement<'a, AnyVecPtr<Traits>>;
+pub type Element<'a, Traits> = ElementPointer<'a, AnyVecPtr<Traits>>;
 
 
 /// Reference to [`Element`].
@@ -129,7 +132,7 @@ pub type Element<'a, Traits> = AnyElement<'a, AnyVecPtr<Traits>>;
 ///
 /// [`AnyVec::get`]: crate::AnyVec::get
 pub struct ElementRef<'a, Traits: ?Sized + Trait = dyn None>(
-    pub(crate) ManuallyDrop<AnyElement<'a, AnyVecPtr<Traits>>>
+    pub(crate) ManuallyDrop<ElementPointer<'a, AnyVecPtr<Traits>>>
 );
 impl<'a, Traits: ?Sized + Trait> Deref for ElementRef<'a, Traits>{
     type Target = Element<'a, Traits>;
@@ -152,7 +155,7 @@ impl<'a, Traits: ?Sized + Trait> Clone for ElementRef<'a, Traits>{
 ///
 /// [`AnyVec::get_mut`]: crate::AnyVec::get_mut
 pub struct ElementMut<'a, Traits: ?Sized + Trait = dyn None>(
-    pub(crate) ManuallyDrop<AnyElement<'a, AnyVecPtr<Traits>>>
+    pub(crate) ManuallyDrop<ElementPointer<'a, AnyVecPtr<Traits>>>
 );
 impl<'a, Traits: ?Sized + Trait> Deref for ElementMut<'a, Traits>{
     type Target = Element<'a, Traits>;