+++ tests/any_vec.rs
@@ -6,7 +6,7 @@ use std::ptr::NonNull;
 use itertools::assert_equal;
 use any_vec::AnyVec;
 use any_vec::any_value::{AnyValueRaw, AnyValueWrapper};
-use any_vec::element::{AnyElement, Element, ElementRef};
+use any_vec::element::{ElementPointer, Element, ElementRef};
 
 #[allow(dead_code)]
 unsafe fn any_as_u8_slice<T: Sized>(p: &T) -> &[u8] {